requirements.txt
django-extensions
django-encrypted-model-fields
djangorestframework
faker
requests
